fix(frontend/backend): UAM (User Access Management) - Groups tm-230 (#619)
* feat(backend): add migrations tm-230

* feat(backend): add group and permission models tm-230

* feat(backend): add group relation to user model tm-230

* feat(backend): add group and permission entities tm-230

* feat(backend): add group and permission repositories tm-230

* style(backend): add empty lines tm-230

* feat(shared): add permission exception and types tm-230

* feat(backend): add permission service tm-230

* feat(backend/shared): add permission controller tm-230

* chore(shared): rename permission-error exception file tm-230

* feat(shared): add permission error message enum tm-230

* feat(shared): add group enum, exception and types tm-230

* feat(backend/shared): update permission repository and service tm-230

* feat(backend): add group relation-name enum and imports from shared tm-230

* feat(backend): update group model, entity and repository tm-230

* feat(backend): add group service tm-230

* feat(backend/shared): add group controller tm-230

* feat(backend/shared): add groups field to users tm-230

* feat(backend): add groups field to other repositories tm-230

* feat(shared): update types, add permission-key and permission-mode enums tm-230

* feat(backend): update entities, repositories and group model tm-230

* feat(backend): update other repositories tm-230

* feat(backend): add check-user-permissions hook tm-230

* feat(backend): add preHandler hook to group and permission controllers tm-230

* refactor(backend): addRoute method tm-230

* chore(backend/shared): add dot and curly braces tm-230

* chore(backend): add constructor type tm-230

* fix(backend): undefined user groups tm-230

* chore(backend): update imports tm-230

* feat(backend/shared): add groupId and permissionId parameters validation tm-230

* feat(backend): update permission service and repository tm-230

* feat(backend): update group controller, service and repository tm-230

* style(shared): add empty line tm-230

* feat(backend): update group service tm-230

* feat(backend): add swagger docs tm-230

* style(backend/shared): update files formatting tm-230

* fix(backend): add groups prop to activity repository tm-230

* fix(backend): add groups prop to activity entity tm-230

* chore(backend): update imports tm-230

* fix(backend): add relations to friend, chat and chat-message queries tm-230

* feat(frontend): add delete, edit and lock icons tm-230

* feat(frontend): update users actions, slice and api tm-230

* feat(frontend): add groups and permissions enums and types tm-230

* feat(frontend): add permissions actions, slice and api tm-230

* feat(frontend): add groups actions, slice and api tm-230

* feat(frontend): update store, add table, table-row and table-cell components tm-230

* feat(frontend): add chip component tm-230

* feat(frontend): add enum, map and groups-tab component tm-230

* feat(frontend): add enum, map and users-tab component tm-230

* feat(frontend): add management page tm-230

* feat(frontend): add management page usage tm-230

* feat(frontend): add edit-modal window tm-230

* feat(frontend): add management slice, update groups, permissions and users slices tm-230

* feat(shared): add all-groups and all-permissions types tm-230

* feat(shared): add group validation schemas tm-230

* feat(frontend/backend/shared): add export from shared tm-230

* feat(backend): update group and permission services and controllers tm-230

* chore(backend): rename parameter tm-230

* feat(frontend): update users actions and api tm-230

* feat(frontend): update groups actions and api tm-230

* feat(frontend): update permissions actions and api tm-230

* feat(frontend): update management slice tm-230

* feat(frontend): update constant, type and enums tm-230

* feat(frontend): add check-if-user-has-permissions helper tm-230

* feat(frontend): add edit-checkbox, update chip and edit-modal components tm-230

* feat(frontend): update table components styles tm-230

* feat(frontend): update table enums and maps tm-230

* feat(frontend): update protected-route and sidebar components tm-230

* feat(frontend): update management page and tabs tm-230

* feat(frontend): add confirmation-modal, update modal and edit-modal tm-230

* feat(frontend): update management slice and groups actions tm-230

* feat(frontend): update users-tab and groups-tab tm-230

* chore(frontend): rename helper folder tm-230

* feat(frontend): add icon style to button component tm-230

* feat(backend/shared): add additional checks to group service tm-230

* feat(frontend): add success notifications tm-230

* feat(frontend): update table and edit-modal components tm-230

* feat(frontend): update groups-tab and rename map tm-230

* feat(frontend): update users-tab and rename map tm-230

* feat(frontend): update management page tm-230

* feat: update database schema tm-230

* fix(backend): relation error on profile page tm-230

* feat(backend): add edit-string-length and insert admin migrations tm-230

* feat(frontend): add group-created notification tm-230

* chore(frontend/backend): quality criteria fixes tm-230

* chore(frontend): revert button styles tm-230

* chore(backend/shared): update error messages tm-230

* chore(frontend): update styles tm-230

* refactor(frontend): code and imports tm-230

* feat(frontend): update groups-tab and users-tab, add constants tm-230

* feat(frontend): remove icon style prop from button component tm-230

* feat(frontend): add user prop to sidebar tm-230

* feat(frontend): update modals tm-230

* feat(frontend): update groups actions and management slice tm-230

* feat(frontend/backend): disable ability to remove uam permission in own group tm-230

* feat(frontend/backend/shared): rename types and update imports tm-230

* feat(frontend): change input-default-value constant tm-230

* fix(backend): update swagger docs tm-230

* fix(frontend): reset checkboxes in modal tm-230

* feat(frontend): add edit-group-modal component tm-230

* feat(frontend): add edit-user-modal component tm-230

* feat(frontend): delete edit-modal component tm-230

* feat(frontend): remove key from edit-checkbox tm-230

* feat(frontend): add default-user and default-group constants tm-230

* refactor(frontend): remove default constants tm-230

---------
