+++ apps/frontend/src/pages/management/libs/components/groups-tab/libs/components/components.ts
@@ -1,0 +1,1 @@
+export { EditGroupModal } from "./edit-group-modal/edit-group-modal.js";
