+++ apps/frontend/src/pages/management/libs/components/groups-tab/styles.module.css
@@ -35,18 +35,6 @@
 	column-gap: 30px;
 }
 
-.modal-list {
-	display: flex;
-	flex-direction: column;
-	row-gap: 20px;
-	padding: 0;
-}
-
-.modal-item {
-	display: flex;
-	column-gap: 10px;
-}
-
 .icon-button {
 	min-height: unset;
 	padding: 0;
