+++ apps/frontend/src/pages/management/libs/components/users-tab/libs/components/components.ts
@@ -1,0 +1,1 @@
+export { EditUserModal } from "./edit-user-modal/edit-user-modal.js";
