+++ apps/frontend/src/pages/management/libs/components/users-tab/libs/components/edit-user-modal/styles.module.css
@@ -1,0 +1,55 @@
+.edit-modal {
+	display: flex;
+	flex-direction: column;
+	flex-grow: 1;
+	justify-content: space-between;
+	max-width: 500px;
+}
+
+.content {
+	display: flex;
+	flex-direction: column;
+	justify-content: space-between;
+	min-height: 300px;
+}
+
+.title {
+	padding: 50px 50px 20px;
+	font-size: var(--font-size-400);
+	font-weight: 500;
+	text-align: center;
+}
+
+.list-container {
+	display: flex;
+	flex-direction: column;
+	align-items: center;
+	max-height: 160px;
+	padding: 0 50px;
+	overflow-y: auto;
+}
+
+.checkbox-list {
+	display: flex;
+	flex-direction: column;
+	row-gap: 20px;
+	padding: 0;
+}
+
+.checkbox-item {
+	display: flex;
+	column-gap: 10px;
+}
+
+.modal-footer {
+	display: flex;
+	justify-content: center;
+	height: 60px;
+	padding: 14px;
+	margin-top: 30px;
+	background-color: var(--color-menu);
+}
+
+.button {
+	width: 80px;
+}
