+++ apps/frontend/src/pages/management/libs/components/users-tab/styles.module.css
@@ -1,15 +1,3 @@
 .table-container {
 	width: 100%;
 }
-
-.modal-list {
-	display: flex;
-	flex-direction: column;
-	row-gap: 20px;
-	padding: 0;
-}
-
-.modal-item {
-	display: flex;
-	column-gap: 10px;
-}
