+++ apps/frontend/src/pages/management/libs/components/users-tab/users-tab.tsx
@@ -21,9 +21,8 @@ import { ManagementDialogueMessage } from "../../enums/enums.js";
 import { ActionsCell } from "../actions-cell/actions-cell.js";
 import { Chip } from "../chip/chip.js";
 import { ConfirmationModal } from "../confirmation-modal/confirmation-modal.js";
-import { EditCheckbox } from "../edit-checkbox/edit-checkbox.js";
-import { EditModal } from "../edit-modal/edit-modal.js";
 import { Table, TableCell, TableRow } from "../table/table.js";
+import { EditUserModal } from "./libs/components/components.js";
 import { USERS_TABLE_HEADERS } from "./libs/constants/constants.js";
 import { UsersTableHeader } from "./libs/enums/enums.js";
 import { usersHeaderToPropertyName } from "./libs/maps/maps.js";
@@ -73,20 +72,6 @@ const UsersTab: React.FC = () => {
 		setIsEditModalOpen(false);
 		setCurrentUser(null);
 	}, []);
-
-	const handleChangeUserGroups = useCallback(
-		(groupId: number, userId: number) => {
-			void dispatch(groupsActions.updateUserGroups({ groupId, userId }));
-		},
-		[dispatch],
-	);
-
-	const handleToggleCheckbox = useCallback(
-		(groupId: number) => {
-			handleChangeUserGroups(groupId, currentUser?.id as number);
-		},
-		[currentUser, handleChangeUserGroups],
-	);
 
 	const [isDeleteModalOpen, setIsDeleteModalOpen] = useState<boolean>(false);
 	const handleOpenDeleteModal = useCallback((user: UserAuthResponseDto) => {
@@ -156,34 +141,15 @@ const UsersTab: React.FC = () => {
 					})}
 				</Table>
 			</div>
-			<EditModal
-				isOpen={isEditModalOpen}
-				onClose={handleCloseEditModal}
-				title={`Edit ${currentUser?.firstName} ${currentUser?.lastName}'s groups:`}
-			>
-				<ul className={styles["modal-list"]}>
-					{groups.map((group) => {
-						const isChecked = Boolean(
-							currentUser?.groups.some((userGroup) => {
-								return userGroup.id === group.id;
-							}),
-						);
-
-						return (
-							<li className={styles["modal-item"]} key={group.id}>
-								<EditCheckbox
-									isChecked={isChecked}
-									itemId={group.id}
-									key={group.id}
-									name={group.name}
-									onToggle={handleToggleCheckbox}
-								/>
-								{group.name}
-							</li>
-						);
-					})}
-				</ul>
-			</EditModal>
+			{currentUser && (
+				<EditUserModal
+					groups={groups}
+					isOpen={isEditModalOpen}
+					onClose={handleCloseEditModal}
+					title={`Edit ${currentUser.firstName} ${currentUser.lastName}'s groups:`}
+					user={currentUser}
+				/>
+			)}
 			<ConfirmationModal
 				isOpen={isDeleteModalOpen}
 				onCancel={handleCloseDeleteModal}
